% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_experiment_infos}
\alias{get_experiment_infos}
\title{Get Fantom's experiment metadata}
\usage{
get_experiment_infos()
}
\value{
A \code{data.frame} with 2 columns: "library_name" and
"sample_name". One row per Fantom experiment.
}
\description{
This function returns the library names and the sample names for every
experiment from Fantom.
}
\examples{
head(get_experiment_infos())
}
\seealso{
get_fantom_library_name
}

