% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_fantom_library_name}
\alias{get_fantom_library_name}
\title{Get a vector of Fantom's libraries ID matching a cell_type}
\usage{
get_fantom_library_name(cell_line)
}
\arguments{
\item{cell_line}{A string corresponding to the name of the cell line.}
}
\value{
A vector with the fantom's libraries ID(s) matching filename
}
\description{
Will check in Fantom's sample names for names that contains the
\code{cell_line} param, ignoring case and will return the results as a
vector of IDs.
}
\examples{
get_fantom_library_name("GM12878")
}

