% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_fantom_tss}
\alias{get_fantom_tss}
\title{Get a GRanges object with all the tss from Fantom.}
\usage{
get_fantom_tss()
}
\value{
A \code{GRanges} object representing every tss from Fantom in
  hg19.
}
\description{
Produce a GRanges object containing all the tss from Fantom but
without the expression values for every experiments.
}
\examples{
get_fantom_tss()
}
\seealso{
\code{\link{get_fantom_tss_tpm}}
}

